Starting to eval distance vs radius.

+++ V1/devices/models.py
@@ -4,7 +4,7 @@ from V1.users.models import User
 class Device(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE, null=True)
     sms_number = models.CharField(max_length=100, null=True)
-    radius = models.IntegerField(default=500)
+    radius = models.FloatField(default=500)
     pin_lat = models.FloatField()
     pin_long = models.FloatField()
